% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisfor.R
\encoding{UTF-8}
\name{multisfor}
\alias{multisfor}
\title{Train a Multi-Source Random Forest Prediction Rule for Prediction using Partly Missing Multi-Omics Data}
\usage{
multisfor(data, folds, num_trees = 500, mtry = NULL, min_node_size = 1)
}
\arguments{
\item{data}{A \code{data.frame} containing the variables in the model.
Must contain a factor-valued column named 'ytarget' with the following
categories '1' (positive class) and '0' (negative class).}

\item{folds}{Vector of integers. Indicating for each row in 'data' to which 
subset it belongs. Observations in the same subset need to feature the same observed 
covariates.}

\item{num_trees}{Number of trees used to construct each forest.}

\item{mtry}{Number of variables to possibly split at in each node. Default is the (rounded down) square root of the number variables.}

\item{min_node_size}{Minimal node size. Default is 1.}
}
\value{
Object of class \code{multisfor}.
}
\description{
Trains a multi-source random forest using multi-omics data that consists of
several subsets which differ according to their missingness patterns.
For example, if their are three data types A, B, C, the first subset may
feature data types A and C, the second subset B and C, the third subset
only A, and the fourth subset A, B, and C.
}
\examples{
\donttest{

# Load some necessary packages:

library(checkmate)
library(randomForestSRC)
library(caret)
library(pROC)
library(doParallel)
library(parallel)
library(ROCR)
library(caret)


# Load the data (not included in the package currently):

load("ExampleData.Rda")


# Fit a multi-source random forest:

fw_rfs <- multisfor(data          = datatrain, 
                    folds         = foldstrain,
                    num_trees     = 10, 
                    mtry          = 50,
                    min_node_size = 1)

}

}
\references{
\itemize{
  \item Breiman, L. (2001). Random forests. Mach Learn, 45:5-32, <\doi{10.1023/A:1010933404324}>.
  }
}
\author{
Roman Hornung, Frederik Ludwigs
}
