% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.multisfor.R
\encoding{UTF-8}
\name{predict.multisfor}
\alias{predict.multisfor}
\title{Predict Using a Fitted Multi-Source Random Forest}
\usage{
\method{predict}{multisfor}(object, data, weighted = TRUE)
}
\arguments{
\item{object}{An object of class 'multisfor'. See \code{\link{multisfor}}.}

\item{data}{A \code{data.frame} containing data for which predictions should
be obtained. Must not contain missing values. Does not have to feature
all variables used to train the multi-source forest, but a subset of these.}

\item{weighted}{Should the predictions of the forests be weighted by their
OOB-AUC values (\code{TRUE}) before averaging them or should an un-weighted average 
be obtained (\code{FALSE}). Default is \code{TRUE}.}
}
\value{
A vector of predicted probabilities for class '1'.
}
\description{
Obtain class probability predictions using a multi-source random forest
fitted using \code{\link{multisfor}}.
}
\examples{
\donttest{

# Load some necessary packages:

library(checkmate)
library(randomForestSRC)
library(caret)
library(pROC)
library(doParallel)
library(parallel)
library(ROCR)
library(caret)


# Load the data (not included in the package currently)

load("ExampleData.Rda")


# Fit a multi-source random forest:

fw_rfs <- multisfor(data          = datatrain, 
                    folds         = foldstrain,
                    num_trees     = 10, 
                    mtry          = 50,
                    min_node_size = 1)


# Get predicted class-probabilites for each obs. in the test-set:

predictions <- predict(object   = fw_rfs, 
                       data     = datatest,
                       weighted = TRUE)

}

}
\references{
\itemize{
  \item Breiman, L. (2001). Random forests. Mach Learn, 45:5-32, <\doi{10.1023/A:1010933404324}>.
  }
}
\author{
Roman Hornung, Frederik Ludwigs
}
